Keep using original monitor data if getting new data fails
Work around EnumDisplayMonitors() not enumerating any monitors while a
UAC prompt is shown by continuing to use the old data in this case: this
might be wrong, but not having any display data is definitely wrong, so
it's still better than the alternative.

The best would be to understand why does EnumDisplayMonitors() fail in
this case and somehow avoid it, but this requires at least reproducing
the problem, which is already challenging.

See https://github.com/wxWidgets/wxWidgets/pull/1281